requirements.txt
Flask==3.1.2
flask-cors==6.0.2
python-dotenv==1.2.1
Werkzeug==3.1.5
google-cloud-vision==3.12.0
google-auth==2.47.0
google-genai==1.47.0
google-ai-generativelanguage==0.6.15
google-api-core==2.29.0
google-api-python-client==2.188.0
google-auth-httplib2==0.3.0
google-generativeai==0.8.6
gunicorn==23.0.0